+++ src/Config.ts
@@ -317,18 +317,6 @@ export interface ConfigParams {
    * @default 20
    *
    * @category Engine
-   *
-   * @deprecated Use {@link binarySearchThreshold} instead.
-   */
-  vlookupThreshold: number,
-  /**
-   * Determines minimum number of elements a range must have in order to use binary search.
-   * Shorter ranges will be searched naively.
-   * Used by VLOOKUP, HLOOKUP and MATCH functions.
-   *
-   * @default 20
-   *
-   * @category Engine
    */
   binarySearchThreshold: number,
   /**
@@ -421,7 +409,6 @@ export class Config implements ConfigParams, ParserConfig {
     precisionRounding: 14,
     useColumnIndex: false,
     useStats: false,
-    vlookupThreshold: 20,
     binarySearchThreshold: 20,
     nullDate: {year: 1899, month: 12, day: 30},
     undoLimit: 20,
@@ -492,8 +479,6 @@ export class Config implements ConfigParams, ParserConfig {
   public readonly useColumnIndex: boolean
   /** @inheritDoc */
   public readonly useStats: boolean
-  /** @inheritDoc */
-  public readonly vlookupThreshold: number
   /** @inheritDoc */
   public readonly binarySearchThreshold: number
   /** @inheritDoc */
@@ -567,7 +552,6 @@ export class Config implements ConfigParams, ParserConfig {
       precisionEpsilon,
       precisionRounding,
       useColumnIndex,
-      vlookupThreshold,
       binarySearchThreshold,
       nullDate,
       useStats,
@@ -611,9 +595,7 @@ export class Config implements ConfigParams, ParserConfig {
     this.validateNumberToBeAtLeast(this.precisionEpsilon, 'precisionEpsilon', 0)
     this.useColumnIndex = this.valueFromParam(useColumnIndex, 'boolean', 'useColumnIndex')
     this.useStats = this.valueFromParam(useStats, 'boolean', 'useStats')
-    this.vlookupThreshold = this.valueFromParam(vlookupThreshold, 'number', 'vlookupThreshold')
-    this.validateNumberToBeAtLeast(this.vlookupThreshold, 'vlookupThreshold', 1)
-    this.binarySearchThreshold = this.valueFromParam(binarySearchThreshold ?? vlookupThreshold, 'number', 'vlookupThreshold')
+    this.binarySearchThreshold = this.valueFromParam(binarySearchThreshold, 'number', 'binarySearchThreshold')
     this.validateNumberToBeAtLeast(this.binarySearchThreshold, 'binarySearchThreshold', 1)
     this.parseDateTime = this.valueFromParam(parseDateTime, 'function', 'parseDateTime')
     this.stringifyDateTime = this.valueFromParam(stringifyDateTime, 'function', 'stringifyDateTime')
@@ -640,8 +622,6 @@ export class Config implements ConfigParams, ParserConfig {
       }
     })
     this.validateNumberToBeAtLeast(this.maxColumns, 'maxColumns', 1)
-
-    this.warnDeprecatedIfUsed(vlookupThreshold, 'vlookupThreshold', 'v.0.3.0', 'binarySearchThreshold')
 
     this.checkIfParametersNotInConflict(
       {value: this.decimalSeparator, name: 'decimalSeparator'},